accordion functionality mainly working
