+++ index.html
@@ -4,6 +4,7 @@
 <head>
   <meta charset="UTF-8" />
   <link rel="icon" type="image/svg+xml" href="/vite.svg" />
+
   <meta name="description" content="Alta Stucco" />
   <meta name="keywords" content="Alta Stucco, Stucco, Exterior Renovation,Residential Services, Commercial Service " />
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
