+++ src/views/contact/contact.tsx
@@ -1,6 +1,7 @@
 import EmbeddedLink from "../../components/embeddedLink/embeddedLink.tsx";
 
 const Contact = () => {
+
     return (
         <>
             <EmbeddedLink />
