+++ src/views/home/home.css
@@ -1,3 +1,13 @@
 .home-container {
     border: 15px solid #000;
 }
+
+.home-container {
+    color: #000000 !important;
+    text-decoration: none !important;
+    background: none !important;
+    border: none !important;
+    padding: 0 !important;
+    font: inherit !important;
+    cursor: pointer !important;
+}
